moved more types not needed in common
